implement discarding optimization for pdcx

+++ src/mpi_util.hpp
@@ -200,13 +200,13 @@ std::vector<DataType> distribute_data_custom(std::vector<DataType>& local_data,
 
 template<typename InputType, typename OutputType>
 std::vector<OutputType> zip_with_index(std::vector<InputType> &input, auto index_function, Communicator<> &comm) {
-    size_t local_size = input.size();
-    const size_t offset = mpi_util::ex_prefix_sum(local_size, comm);
+    uint64_t local_size = input.size();
+    const uint64_t offset = mpi_util::ex_prefix_sum(local_size, comm);
 
     std::vector<OutputType> zipped;
     zipped.reserve(local_size);
-    for (size_t i = 0; i < local_size; ++i) {
-        size_t index = offset + i;
+    for (uint64_t i = 0; i < local_size; ++i) {
+        uint64_t index = offset + i;
         zipped.push_back(index_function(index, input[i]));
     }
     return zipped;
